separate functions out of cobra runCmd
